Moved textline merge to detection folder and started transition to TextBlocks
