+++ ocr/model_32px.py
@@ -11,6 +11,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
+from detection.ctd_utils.textblock import TextBlock
 from utils import Quadrilateral, chunks
 from .common import OfflineOCR
 
@@ -44,7 +45,7 @@ class Model32pxOCR(OfflineOCR):
     async def _unload(self):
         del self.model
     
-    async def _forward(self, image: np.ndarray, textlines: List[Quadrilateral], verbose: bool = False) -> List[Quadrilateral]:
+    async def _forward(self, image: np.ndarray, textlines: List[TextBlock], verbose: bool = False) -> List[TextBlock]:
         text_height = 32
         max_chunk_size = 16
